MetaUiHandler and drawBox own files
